Move functions for sleep tracker view to tracker view for displaying tracking data #672 #686
